refactor(MmuPanel): extract climate info into MmuUnitFooterClimate component
Move the per-gate climate, heater and drying display out of MmuUnitFooter
into a dedicated MmuUnitFooterClimate component. Render the per-gate report
as structured rows via v-for instead of a pre-line string, and drop the
manual tooltip-open watch in favor of a computed property.
